+++ src/openapi/transport/batch.js
@@ -67,20 +67,8 @@ function batchCallSuccess(callList, batchResult) {
 function runBatchCall(serviceGroup, callList) {
 
     const subRequests = [];
-    let authToken;
     for (let i = 0, call; call = callList[i]; i++) {
-        let headers = call.options && call.options.headers;
-        if (headers && headers.Authorization) {
-            authToken = headers.Authorization;
-
-            const newHeaders = {};
-            for (const header in headers) {
-                if (header !== 'Authorization') {
-                    newHeaders[header] = headers[header];
-                }
-            }
-            headers = newHeaders;
-        }
+        const headers = call.options && call.options.headers;
         let body = call.options && call.options.body;
         if (typeof body !== 'string') {
             body = JSON.stringify(body);
@@ -93,12 +81,8 @@ function runBatchCall(serviceGroup, callList) {
         });
     }
 
-    if (!authToken) {
-        authToken = this.authProvider.getToken();
-    }
-
     const boundary = '~';
-    const content = buildBatch(subRequests, boundary, authToken, this.host);
+    const content = buildBatch(subRequests, boundary, this.host);
 
     this.transport.post(serviceGroup, 'batch', null, {
         headers: { 'Content-Type': 'multipart/mixed; boundary="' + boundary + '"' },
@@ -118,13 +102,11 @@ function runBatchCall(serviceGroup, callList) {
  * @alias saxo.openapi.TransportBatch
  * @param {Transport} transport - Instance of the transport class to wrap.
  * @param {string} baseUrl - Base URL for batch requests. This should be an absolute URL.
- * @param {{getToken:function}} [authProvider] - Optional instance of an auth provider, such as TransportAuth.auth, used to add
- *      authentication to each batch item.
  * @param {Object} [options]
  * @param {number} [options.timeoutMs=0] - Timeout after starting to que items before sending a batch request.
  * @param {string} [options.host=global.location.host] - The host to use in the batch request. If not set defaults to global.location.host.
  */
-function TransportBatch(transport, baseUrl, authProvider, options) {
+function TransportBatch(transport, baseUrl, options) {
     TransportQueue.call(this, transport);
 
     if (!baseUrl) {
@@ -152,7 +134,6 @@ function TransportBatch(transport, baseUrl, authProvider, options) {
         this.host = location.host;
     }
 
-    this.authProvider = authProvider;
     this.timeoutMs = options && options.timeoutMs || 0;
     this.isQueueing = true;
 }